app-viewlib: rename mur-editor-mgr to mur-setting-client
